requirements.txt
django
psycopg2
gunicorn
django-heroku
python-gettext
djangorestframework
